✨ Add settings
