+++ src/app/components/icons/MenuIcon.tsx
@@ -1,0 +1,9 @@
+function MenuIcon(): JSX.Element {
+  return (
+    <svg height="24px" viewBox="0 0 24 24" width="24px" fill="currentColor">
+      <path d="M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z" />
+    </svg>
+  );
+}
+
+export default MenuIcon;
