+++ src/app/sender/Settings.module.css
@@ -1,0 +1,34 @@
+.container {
+  position: absolute;
+  inset: 0;
+  z-index: 1;
+  background: var(--color-shade-one);
+  display: grid;
+}
+
+.label {
+  display: grid;
+  grid-template-columns: 50% 50%;
+  align-items: center;
+}
+
+.close {
+  position: absolute;
+  right: 0.2em;
+  top: 0.2em;
+  z-index: 2;
+  width: 2em;
+  height: 2em;
+}
+
+.settings {
+  overflow: auto;
+  display: grid;
+  row-gap: 1rem;
+}
+
+.logout {
+  background: var(--color-shade-two);
+  color: lightcoral;
+  padding: 0.5em;
+}
